+++ app/model/CreateCollectionData.scala
@@ -1,3 +1,5 @@
 package model
 
 case class CreateCollectionData(collectionName: String)
+
+case class ConsignmentForm(name: String)
