+++ app/views/consignments.scala.html
@@ -1,0 +1,58 @@
+@import forms.ConsignmentForm
+@(userForm: Form[ConsignmentForm])(implicit messages: Messages)
+    @import helper._
+    @main("Consignment") {
+        @defining(play.core.PlayVersion.current) { version =>
+
+            <!-- Back -->
+            <a href="@routes.SeriesDetailsController.index()" class="govuk-back-link">Back</a>
+
+            <!-- Progress indicator -->
+            <span class="govuk-caption-xl">Step 2 of 5</span>
+            <!-- Headline -->
+            <h1 class="govuk-heading-xl">Consignment Information</h1>
+
+            <div class="govuk-grid-row">
+                <div class="govuk-grid-column-two-thirds">
+                @form(action = routes.CreateConsignmentController.submit) {
+                    <fieldset class="govuk-fieldset">
+
+                            <!-- Consignment Name -->
+                        <div class="govuk-form-group">
+                            <label class="govuk-label">
+                                Consignment Name*
+                            </label>
+                            @inputText(userForm("consignmentName"),
+                                '_label -> null,
+                                'class -> "govuk-input",
+                                '_id -> "consignmentName")
+                        </div>
+
+                            <!-- Transferring Body -->
+                        <div class="govuk-form-group">
+                            <label class="govuk-label">
+                                Transferring Body*
+                            </label>
+                            @inputText(userForm("transferringBody"),
+                                '_label -> null,
+                                'class -> "govuk-input")
+                            <input type="hidden" name="series" value="1" />
+                        </div>
+
+                    </fieldset>
+
+                        <!-- Buttons -->
+                    <div>
+                            <!-- Cancel -->
+                        <a href="@routes.DashboardController.index" role="button" draggable="false" class="govuk-button govuk-button--secondary" data-module="govuk-button">
+                            Cancel
+                        </a>
+                            <!-- Continue -->
+                        <input type="submit" value="Save and continue" role="button" draggable="false" class="govuk-button" data-module="govuk-button"/>
+                    </div>
+
+                }
+                </div>
+            </div>
+        }
+    }
