--- app/views/createCollection.scala.html
@@ -1,38 +1,0 @@
-@(userForm: Form[model.CreateCollectionData], seriesId: Int)
-
-@main("CreateCollection") {
-@defining(play.core.PlayVersion.current) { version =>
-
-<div className="govuk-grid-row">
-    <div className="govuk-grid-column-two-thirds">
-        <h1 class="govuk-heading-xl">Consignment Information</h1>
-        <form
-                action="@routes.CreateCollectionController.submit()"
-                method="POST"
-        >
-
-            <div className="govuk-form-group">
-                <label className="govuk-label" htmlFor={props.id}>
-                    Consignment Name
-                </label>
-                <input
-                        className={`govuk-input govuk-!-width-one-third`}
-                        id="consignmentName"
-                        name="consignmetName"
-                />
-            </div>
-            <div class="govuk-form-group">
-                <button class="govuk-button govuk-button--secondary" type="submit" data-module="govuk-button" role="button">
-                    Cancel
-                </button>
-                <button class="govuk-button" type="submit" data-module="govuk-button" role="button">
-                    Save and continue
-                </button>
-            </div>
-
-        </form>
-    </div>
-</div>
-
-}
-}
