--- app/views/showCollections.scala.html
@@ -1,20 +1,0 @@
-@import graphql.codegen.GetConsignments.getConsignments.GetConsignments
-@(collections:Seq[GetConsignments])
-
-@main("ViewCollections") {
-@defining(play.core.PlayVersion.current) { version =>
-<h1 class="govuk-heading-xl">View Batches</h1>
-
-<div>
-    <ul>
-        @for(consignment <- collections) {
-            <li>
-                <p>Consignment Name: @consignment.name</p>
-                <p>Consignment Id : @consignment.id</p>
-                <p>Series: @consignment.series.description</p> }
-            </li>
-    </ul>
-</div>
-
-}
-}
