+++ conf/application.conf
@@ -21,7 +21,7 @@ aws.secret.access.key = ${?AWS_SECRET_ACCESS_KEY}
 aws.access.key.id = ${?AWS_ACCESS_KEY_ID}
 
 graphql.uri = "http://localhost:8080/graphql"
-graphql.uri = ${?TDR_GRAPHQL_URI}
+//graphql.uri = ${?TDR_GRAPHQL_URI}
