+++ conf/routes
@@ -13,11 +13,14 @@ POST    /createSeries/submit                controllers.CreateSeriesController.s
 GET     /upload                             controllers.UploadController.index(consignmentId: Int)
 GET     /authenticate/cognito               controllers.AuthController.authenticate
 GET     /viewCollections                    controllers.ViewCollectionsController.index
-GET     /createCollection                   controllers.CreateCollectionController.index(seriesId: Int)
-POST    /createCollection/submit            controllers.CreateCollectionController.submit()
+GET     /createConsignment                   controllers.CreateConsignmentController.index(seriesId: Int)
+POST    /createConsignment/submit            controllers.CreateConsignmentController.submit()
 GET     /fileStatus                         controllers.FileStatusController.getFileStatus(consignmentId: Int)
 GET    /fileStatusApi                       controllers.FileStatusController.getFileStatusApi(consignmentId: Int)
 POST    /upload                             controllers.UploadController.upload(consignmentId: Int)
 POST    /filedata                           controllers.UploadController.saveFileData
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
+GET     /authenticate/cognito               controllers.AuthController.authenticate
+
+
